% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/box2upper.R
\name{box2upper}
\alias{box2upper}
\title{Returns the upper bound value for a given boxnumber}
\usage{
box2upper(boxnumber, boxsize = 1, log = FALSE)
}
\arguments{
\item{boxnumber}{An integer boxnumber}

\item{boxsize}{single numeric value, used as the boxsize}

\item{log}{TRUE, if logarithmic scales should be used}
}
\description{
Returns the upper bound value for a given boxnumber
}

