% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bp.signalprocessor.R
\name{bp.signalprocessor}
\alias{bp.signalprocessor}
\title{This function identifies chart signals in an [0,100]-Points Bullish Percent Chart}
\usage{
bp.signalprocessor(data)
}
\arguments{
\item{data}{Input data}
}
\description{
This function identifies chart signals in an [0,100]-Points Bullish Percent Chart
}

