\name{CalcPerformanceMap}
\alias{CalcPerformanceMap}
\title{This function calculates an upper triangle performance map for a given time
series.}
\usage{
  CalcPerformanceMap(quotes, column = "close",
    mode = c("abs", "rel"))
}
\description{
  This function calculates an upper triangle performance
  map for a given time series.
}

