\name{currentVPOBreakoutMethod}
\alias{currentVPOBreakoutMethod}
\title{Identifiy for a given P&F Table the current vertical price objective
triggered by the last signal reversal.}
\usage{
  currentVPOBreakoutMethod(data, reversal, boxsize, log)
}
\description{
  Identifiy for a given P&F Table the current vertical
  price objective triggered by the last signal reversal.
}

