\name{currentVPOReversalMethod}
\alias{currentVPOReversalMethod}
\title{Identifiy for a given P&F Table the current vertical price objective
triggered by the last signal reversal.}
\usage{
  currentVPOReversalMethod(data, reversal, boxsize, log)
}
\description{
  Identifiy for a given P&F Table the current vertical
  price objective triggered by the last signal reversal.
}

