\name{doubleBottom}
\alias{doubleBottom}
\title{returns true if given column c matches exactly previous column of same type (this is always column c-2)}
\usage{
  doubleBottom(redData, column)
}
\description{
  returns true if given column c matches exactly previous
  column of same type (this is always column c-2)
}

