\name{doubleTop}
\alias{doubleTop}
\title{Returns true if given column c matches exactly previous column of same type (this is always column c-2)}
\usage{
  doubleTop(redData, column)
}
\description{
  Returns true if given column c matches exactly previous
  column of same type (this is always column c-2)
}

