% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rpnf-package.R
\docType{data}
\name{DOW}
\alias{DOW}
\title{This is some free available quote data for the DOW Chemical Company.}
\format{\preformatted{'data.frame':	248 obs. of  13 variables:
 $ Date       : Date, format: "2015-10-09" "2015-10-08" "2015-10-07" "2015-10-06" ...
 $ Open       : num  48 47.3 47.4 46.5 44.6 ...
 $ High       : num  48.1 48.1 47.9 47.8 46.5 ...
 $ Low        : num  46.8 47.3 46.6 46.2 44.6 ...
 $ Close      : num  47.1 47.8 47.5 47.1 46.3 ...
 $ Volume     : num  13662925 9522229 30866454 18611813 79471630 ...
 $ Ex-Dividend: num  0 0 0 0 0 0 0 0 0 0.42 ...
 $ Split Ratio: num  1 1 1 1 1 1 1 1 1 1 ...
 $ Adj. Open  : num  48 47.3 47.4 46.5 44.6 ...
 $ Adj. High  : num  48.1 48.1 47.9 47.8 46.5 ...
 $ Adj. Low   : num  46.8 47.3 46.6 46.2 44.6 ...
 $ Adj. Close : num  47.1 47.8 47.5 47.1 46.3 ...
 $ Adj. Volume: num  13662925 9522229 30866454 18611813 79471630 ...
 - attr(*, "freq")= chr "daily"
}}
\usage{
DOW
}
\description{
End of day open, high, low, close and volume, dividends and splits,
 and split/dividend adjusted open, high, low close and volume for Dow Chemical Company (The) (DOW).
 Data are freely available at https://www.quandl.com/data/WIKI/DOW,
 and may be copy, distribute, disseminate or include the data in other products for commercial and/or noncommercial purposes.
 This data is part of Quandl's Wiki initiative to get financial data permanently into the public domain.
 Quandl relies on users like you to flag errors and provide data where data is wrong or missing.
 Get involved: connect@quandl.com
}
\author{
Sascha Herrmann \email{sascha.herrmann.consulting@gmail.com}
}
\references{
\url{https://www.quandl.com/data/WIKI/DOW}
}
\keyword{data}

