\name{fallingBottom}
\alias{fallingBottom}
\title{Returns true if given column c drops below prevois column of same type (this is always column c-2)}
\usage{
  fallingBottom(redData, column)
}
\description{
  Returns true if given column c drops below prevois column
  of same type (this is always column c-2)
}

