\name{fallingTop}
\alias{fallingTop}
\title{returns true if given column c drops below previous column of same type (this is always column c-2)}
\usage{
  fallingTop(redData, column)
}
\description{
  returns true if given column c drops below previous
  column of same type (this is always column c-2)
}

