% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/maxBox.R
\name{maxBox}
\alias{maxBox}
\title{Returns the maximum box number in given column}
\usage{
maxBox(redData, column)
}
\arguments{
\item{redData}{Data to consider}

\item{column}{Column to consider}
}
\description{
Returns the maximum box number in given column
}

