% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/minBox.R
\name{minBox}
\alias{minBox}
\title{Returns the minimum box number in given column}
\usage{
minBox(redData, column)
}
\arguments{
\item{redData}{Data to consider}

\item{column}{Column to consider}
}
\description{
Returns the minimum box number in given column
}

