% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextBox.R
\name{nextBox}
\alias{nextBox}
\title{Determine the next box frontier for current quote(s) given a recent XO-status.}
\usage{
nextBox(quote, status, boxsize = 1, log = FALSE)
}
\arguments{
\item{quote}{A single quote or a vector of quotes.}

\item{status}{A single character indicating the current XO-status.}

\item{boxsize}{A single numeric value, indicating the boxsize to be considered.}

\item{log}{TRUE, if logarithmic scales should be used.}
}
\description{
Note: offset should only be used for reversal calculation
}

