\name{PlotPerformanceMap}
\alias{PlotPerformanceMap}
\title{This function plots a performance map as an heatmap.}
\usage{
  PlotPerformanceMap(quotes, column = "close",
    rescaling = "binary", ...)
}
\description{
  TODO: Create a proper colormap centered at 0 (perhaps via
  rescaling) TODO: Label axes with date instead of numbers
}
\seealso{
  \url{http://www.nytimes.com/interactive/2011/01/02/business/20110102-metrics-graphic.html}
}

