\name{pnfplot}
\alias{pnfplot}
\title{Generate a point and figure plot}
\usage{
  pnfplot(data, ...)
}
\arguments{
  \item{data}{a data frame object containing point and
  figure informations to be plotted}

  \item{...}{any additional options for the plot command}
}
\description{
  Generate a point and figure plot
}

