% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/quote2box.R
\name{quote2box}
\alias{quote2box}
\title{Converts a single or a vector of quotes into integer boxnumbers for P&F-Analysis.}
\usage{
quote2box(quote, boxsize = 1, log = FALSE)
}
\arguments{
\item{quote}{a single quote, or a vector of quotes}

\item{boxsize}{single numeric value, used as the boxsize}

\item{log}{TRUE, if logarithmic scales should be used}
}
\value{
a single or a vector of integer boxnumbers
This function transforms a given quote into an unique integer box number
}
\description{
Converts a single or a vector of quotes into integer boxnumbers for P&F-Analysis.
}

