% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoteToBoxnumber.R
\name{quoteToBoxnumber}
\alias{quoteToBoxnumber}
\title{Determines the boxnumber for a given tuple of quotes,status,boxsize and log.}
\usage{
quoteToBoxnumber(quote, status, boxsize, log)
}
\arguments{
\item{quote}{a numeric vector of quotes}

\item{status}{current status, either "X" or "O"}

\item{boxsize}{boxsize}

\item{log}{use log scale, either TRUE or FALSE}
}
\value{
a vector of integer boxnumbers
}
\description{
Determines the boxnumber for a given tuple of quotes,status,boxsize and log.
}

