% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quoteToScale.R
\name{quoteToScale}
\alias{quoteToScale}
\title{Scales a quote. In case log==TRUE this is logarithmic scale, original scale otherwise.}
\usage{
quoteToScale(x, log)
}
\arguments{
\item{x}{a numeric vector of quotes}

\item{log}{TRUE or FALSE}
}
\value{
scaled quote
}
\description{
Scales a quote. In case log==TRUE this is logarithmic scale, original scale otherwise.
}

