\name{raisingBottom}
\alias{raisingBottom}
\title{returns true if given column c exceeds prevois column of same type (this is always column c-2)}
\usage{
  raisingBottom(redData, column)
}
\description{
  returns true if given column c exceeds prevois column of
  same type (this is always column c-2)
}

