\name{raisingTop}
\alias{raisingTop}
\title{Returns true if given column c exceeds previous column of same type (this is always column c-2)}
\usage{
  raisingTop(redData, column)
}
\description{
  Returns true if given column c exceeds previous column of
  same type (this is always column c-2)
}

