% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rs.signal.processor.R
\name{rs.signal.processor}
\alias{rs.signal.processor}
\title{This function analyzes a (preliminary) P&F Chart for Bullish Support Line and Bearish Resistance Line}
\usage{
rs.signal.processor(data)
}
\arguments{
\item{data}{Input data}
}
\description{
Finding the appropriate trendlines is explained very good at \url{http://stockcharts.com/school/doku.php?id=chart_school:chart_analysis:pnf_charts:pnf_trendlines}.
}
\seealso{
\url{http://stockcharts.com/school/doku.php?id=chart_school:chart_analysis:pnf_charts:pnf_trendlines}
}

