% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleToQuote.R
\name{scaleToQuote}
\alias{scaleToQuote}
\title{Rescales a scaled quote to original scale.}
\usage{
scaleToQuote(x, log)
}
\arguments{
\item{x}{a numeric vector of scaled quotes}

\item{log}{TRUE or FALSE}
}
\value{
scaled quote
}
\description{
Rescales a scaled quote to original scale.
}

