% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/signalanalyzer.R
\name{signalanalyzer}
\alias{signalanalyzer}
\title{analyze transitions of signal states}
\usage{
signalanalyzer(signal, probability = TRUE)
}
\arguments{
\item{signal}{Signal to identify}

\item{probability}{Report probability}
}
\description{
analyze transitions of signal states
}

