\name{xo.priceobjective.processor}
\alias{xo.priceobjective.processor}
\title{This function analyzes a (preliminary) P&F Chart for Price Objectives}
\usage{
  xo.priceobjective.processor(data)
}
\description{
  Finding the appropriate price objectives is explained
  very good at
  \url{http://stockcharts.com/school/doku.php?id=chart_school:chart_analysis:point_and_figure_pri}.
}
\seealso{
  \url{http://stockcharts.com/school/doku.php?id=chart_school:chart_analysis:point_and_figure_pri}
}

