% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xo.signalprocessor.R
\name{xo.signalprocessor}
\alias{xo.signalprocessor}
\title{Analyzes a given PNF time-series for Buy&Sell patterns}
\usage{
xo.signalprocessor(data, reversal = 3)
}
\arguments{
\item{data}{Input data}

\item{reversal}{Number of boxes for reversal}
}
\description{
Analyzes a given PNF time-series for Buy&Sell patterns
}

