\name{xo.trendline.processor}
\alias{xo.trendline.processor}
\title{This function analyzes a (preliminary) P&F Chart for Bullish Support Line and Bearish Resistance Line}
\usage{
  xo.trendline.processor(data, slope = 1)
}
\description{
  Finding the appropriate trendlines is explained very good
  at
  \url{http://stockcharts.com/school/doku.php?id=chart_school:chart_analysis:pnf_charts:pnf_trendlines}.
}
\seealso{
  \url{http://stockcharts.com/school/doku.php?id=chart_school:chart_analysis:pnf_charts:pnf_trendlines}
}

