% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/box2lower.R
\name{box2lower}
\alias{box2lower}
\title{Returns the lower bound value for a given boxnumber}
\usage{
box2lower(boxnumber, boxsize = 1, log = FALSE)
}
\arguments{
\item{boxnumber}{An integer boxnumber}

\item{boxsize}{single numeric value, used as the boxsize}

\item{log}{TRUE, if logarithmic scales should be used}
}
\description{
Returns the lower bound value for a given boxnumber
}

