% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/currentVPOBreakoutMethod.R
\name{currentVPOBreakoutMethod}
\alias{currentVPOBreakoutMethod}
\title{Identifiy for a given P&F Table the current vertical price objective 
triggered by the last signal reversal.}
\usage{
currentVPOBreakoutMethod(data, reversal, boxsize, log)
}
\arguments{
\item{data}{Input data}

\item{reversal}{Number of boxes for reversal}

\item{boxsize}{Size of one box}

\item{log}{Use logarithmic scale}
}
\description{
Identifiy for a given P&F Table the current vertical price objective 
triggered by the last signal reversal.
}

