% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/doubleBottom.R
\name{doubleBottom}
\alias{doubleBottom}
\title{returns true if given column c matches exactly previous column of same type (this is always column c-2)}
\usage{
doubleBottom(redData, column)
}
\arguments{
\item{redData}{Data to consider}

\item{column}{Column to consider}
}
\description{
returns true if given column c matches exactly previous column of same type (this is always column c-2)
}

