% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpnf-package.R
\docType{data}
\name{DOW}
\alias{DOW}
\title{This is some free available quote data for the DOW Chemical Company.}
\description{
End of day open, high, low, close and volume, dividends and splits, 
 and split/dividend adjusted open, high, low close and volume for Dow Chemical Company (The) (DOW).  
 Data are freely available at https://www.quandl.com/data/WIKI/DOW,
 and may be copy, distribute, disseminate or include the data in other products for commercial and/or noncommercial purposes. 
 This data is part of Quandl's Wiki initiative to get financial data permanently into the public domain. 
 Quandl relies on users like you to flag errors and provide data where data is wrong or missing. 
 Get involved: connect@quandl.com
}
\author{
Sascha Herrmann \email{sascha.herrmann.consulting@gmail.com}
}
\references{
\url{https://www.quandl.com/data/WIKI/DOW}
}
\keyword{data}

