% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fallingBottom.R
\name{fallingBottom}
\alias{fallingBottom}
\title{Returns true if given column c drops below prevois column of same type (this is always column c-2)}
\usage{
fallingBottom(redData, column)
}
\arguments{
\item{redData}{Data to consider}

\item{column}{Column to consider}
}
\description{
Returns true if given column c drops below prevois column of same type (this is always column c-2)
}

