% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextReversal.R
\name{nextReversal}
\alias{nextReversal}
\title{Determine the next reversal frontier for current quote(s) given a recent XO-status.}
\usage{
nextReversal(quote, status, reversal = 3L, boxsize = 1, log = FALSE)
}
\arguments{
\item{quote}{A single quote or a vector of quotes.}

\item{status}{A single character indicating the current XO-status.}

\item{reversal}{number of boxes needed to make a reversal}

\item{boxsize}{A single numeric value, indicating the boxsize to be considered.}

\item{log}{TRUE, if logarithmic scales should be used.}
}
\description{
Determine the next reversal frontier for current quote(s) given a recent XO-status.
}

