\name{pnfplot}
\alias{pnfplot}
\title{Generate a modern point and figure plot}
\usage{
  pnfplot(data, ...)
}
\arguments{
  \item{data}{a data frame object containing point and
  figure informations to be plotted}

  \item{...}{any additional options for the plot command}
}
\description{
  Generate a modern point and figure plot
}
\seealso{
  \code{\link{pnfprocessor}}

  \code{\link{pnfplottxt}}
}

