% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pnfplottxt.R
\name{pnfplottxt}
\alias{pnfplottxt}
\title{Generate a classical TXT point and figure plot.}
\usage{
pnfplottxt(data, reversal = 3, boxsize = 1, log = FALSE, main = NULL,
  sub = NULL)
}
\arguments{
\item{data}{a data frame object containing point and figure informations to be plotted}

\item{reversal}{number of boxes used in pnfprocessor}

\item{boxsize}{the boxsize used in pnfprocessor}

\item{log}{are calculations done in logarithmic mode}

\item{main}{a string used as a main title of the chart}

\item{sub}{a string used as a sub title of the chart}
}
\description{
THIS FUNCTION IS STILL UNDER DEVELOPMENT,
THEREFORE IT MIGHT BE SUBJECT TO CHANGE!
}
\examples{
library(rpnf) # Load rpnf library
data(DOW) # (Offline) Load free available sample data from https://www.quandl.com/data/WIKI/DOW
pnfdata <- pnfprocessor(
  high=DOW$High,
  low=DOW$Low,
  date=DOW$Date,
  boxsize=1L,
  log=FALSE)
pnfplottxt(pnfdata,boxsize=1L,log=FALSE)
}
\references{
\url{http://rpnf.r-forge.r-project.org}
}
\seealso{
\code{\link{pnfprocessor}}

\code{\link{pnfplot}}
}

