% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raisingBottom.R
\name{raisingBottom}
\alias{raisingBottom}
\title{returns true if given column c exceeds prevois column of same type (this is always column c-2)}
\usage{
raisingBottom(redData, column)
}
\arguments{
\item{redData}{Data to consider}

\item{column}{Column to consider}
}
\description{
returns true if given column c exceeds prevois column of same type (this is always column c-2)
}

