% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raisingTop.R
\name{raisingTop}
\alias{raisingTop}
\title{Returns true if given column c exceeds previous column of same type (this is always column c-2)}
\usage{
raisingTop(redData, column)
}
\arguments{
\item{redData}{Data to consider}

\item{column}{Column to consider}
}
\description{
Returns true if given column c exceeds previous column of same type (this is always column c-2)
}

