% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xo.priceobjective.processor.R
\name{xo.priceobjective.processor}
\alias{xo.priceobjective.processor}
\title{This function adds Vertical Price Objectives calculated with the
Bullish Breakout and Bearish Breakdown Method (BM) to an P&F Table.}
\usage{
xo.priceobjective.processor(data, reversal, boxsize, log)
}
\arguments{
\item{data}{Input data}

\item{reversal}{Number of boxes for reversal}

\item{boxsize}{Size of one box}

\item{log}{Use logarithmic scale}
}
\description{
Finding the appropriate price objectives has been explained very good at
http://stockcharts.com/school/doku.php?id=chart_school:chart_analysis:point_and_figure_pri,
but this documentation is no longer available.
The function adds columns vpo_bm_boxnumber and vpo_bm_price to the given
P&F Table. vpo_bm_bonumber contains the boxnumber of the price objective,
while vpo_bm_price contains the real price objective.
}

