% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xo.processor.R
\name{xo.processor}
\alias{xo.processor}
\title{Determine the XO development of a given time series.}
\usage{
xo.processor(high, low = high, date, reversal = 3L, boxsize = 1,
  log = FALSE)
}
\arguments{
\item{high}{Vector of high quotes}

\item{low}{Vector of low quotes. If skipped, high are taken as low.}

\item{date}{Vector of dates of the time series.}

\item{reversal}{Number of boxes necessary for a reversal}

\item{boxsize}{A single numeric value, indicating the boxsize to be considered.}

\item{log}{TRUE, if logarithmic scales should be used.}
}
\description{
This is the main PNF-Workhorse, which transforms a given time series into X and O's.
Furthermore it provides already simple Buy/Sell signals, based on checking if the second last colum.
}
\keyword{internal}

