% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/xo.processor.R
\name{xo.processor}
\alias{xo.processor}
\title{Determine the XO development of a given time series.}
\usage{
xo.processor(high, low = high, date, reversal = 3L, boxsize = 1,
  log = FALSE)
}
\arguments{
\item{high}{}

\item{low}{}

\item{date}{}

\item{reversal}{}

\item{boxsize}{A single numeric value, indicating the boxsize to be considered.}

\item{log}{TRUE, if logarithmic scales should be used.}
}
\description{
This is the main PNF-Workhorse, which transforms a given time series into X and O's.
Furthermore it provides already simple Buy/Sell signals, based on checking if the second last colum.
}
\keyword{internal}

